+++ src/test/java/io/personium/common/auth/token/ResidentRefreshTokenTest.java
@@ -1,6 +1,8 @@
 /**
- * personium.io
- * Copyright 2014 FUJITSU LIMITED
+ * Personium
+ * Copyright 2019 Personium Project Autors
+ *  - Fujitsu Ltd.
+ *  - Akio Shimono
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.