+++ CHANGELOG.md
@@ -1,5 +1,10 @@
+## 1.5.3
+BREAKING CHANGES:
+* Bytes used for CBC Initial Vector in Local Token encryption changed. ([#73](https://github.com/personium/personium-lib-common/issues/73))
+
+
 ## 1.5.2
 BREAKING CHANGES:
-* all classes are put under io.personium.common.*
+* All classes are put under io.personium.common.*